git_mulagent/langchain_agent_transformer.py
from typing import Dict, List, Optional
from langchain.agents import AgentExecutor, create_structured_chat_agent
from langchain_core.messages import HumanMessage, SystemMessage
from langchain_openai import ChatOpenAI
from langchain.tools import Tool
from langchain.memory import ConversationBufferMemory
from langchain.prompts import ChatPromptTemplate, MessagesPlaceholder
import json
import os

# 设置API密钥
API_KEY = "sk-15ece1d22cf2433fa0dd4fcc874b6154"

class CodeAnalysisTool:
    def __init__(self):
        self.name = "code_analyzer"
        self.description = "分析代码结构，识别关键组件"
        
    def run(self, code: str) -> Dict:
        # 实现代码分析逻辑
        return {
            "imports": [],
            "globals": [],
            "classes": [],
            "functions": [],
            "main": None
        }

class ImportConverterTool:
    def __init__(self):
        self.name = "import_converter"
        self.description = "转换导入语句，处理依赖关系"
        
    def run(self, imports: List[str], rules: Dict) -> str:
        # 实现导入转换逻辑
        return ""

class ClassConverterTool:
    def __init__(self):
        self.name = "class_converter"
        self.description = "转换类定义，处理继承关系"
        
    def run(self, classes: List[Dict], rules: Dict) -> str:
        # 实现类转换逻辑
        return ""

class FunctionConverterTool:
    def __init__(self):
        self.name = "function_converter"
        self.description = "转换函数定义，处理参数和返回值"
        
    def run(self, functions: List[Dict], rules: Dict) -> str:
        # 实现函数转换逻辑
        return ""

class MainConverterTool:
    def __init__(self):
        self.name = "main_converter"
        self.description = "转换主程序入口，处理程序流程"
        
    def run(self, main_code: str, rules: Dict) -> str:
        # 实现主程序转换逻辑
        return ""

class CodeIntegratorTool:
    def __init__(self):
        self.name = "code_integrator"
        self.description = "整合所有转换后的代码，确保一致性"
        
    def run(self, parts: Dict) -> str:
        # 实现代码整合逻辑
        return ""

class LangChainTransformer:
    def __init__(self):
        # 配置DeepSeek
        self.llm = ChatOpenAI(
            temperature=0,
            model="deepseek-chat",
            openai_api_key=API_KEY,
            openai_api_base="https://api.deepseek.com/v1",
            model_kwargs={"openai_api_type": "deepseek"}
        )
        
        # 配置记忆
        self.memory = ConversationBufferMemory(
            memory_key="chat_history",
            return_messages=True
        )
        
        # 配置工具
        self.tools = [
            Tool(
                name="code_analyzer",
                func=CodeAnalysisTool().run,
                description="分析代码结构，识别关键组件"
            ),
            Tool(
                name="import_converter",
                func=ImportConverterTool().run,
                description="转换导入语句，处理依赖关系"
            ),
            Tool(
                name="class_converter",
                func=ClassConverterTool().run,
                description="转换类定义，处理继承关系"
            ),
            Tool(
                name="function_converter",
                func=FunctionConverterTool().run,
                description="转换函数定义，处理参数和返回值"
            ),
            Tool(
                name="main_converter",
                func=MainConverterTool().run,
                description="转换主程序入口，处理程序流程"
            ),
            Tool(
                name="code_integrator",
                func=CodeIntegratorTool().run,
                description="整合所有转换后的代码，确保一致性"
            )
        ]
        
        # 配置提示词模板
        self.prompt = ChatPromptTemplate.from_messages([
            ("system", """你是一个代码转换专家，负责将代码从一种编程语言转换为另一种。
            请按照以下步骤进行转换：
            1. 分析代码结构
            2. 转换各个组件
            3. 整合最终代码
            
            可用的工具：
            {tools}
            
            转换规则：
            {rules}
            
            请使用以下格式：
            思考：我需要做什么
            行动：使用哪个工具
            行动输入：工具的输入
            观察：工具的输出
            思考：下一步该做什么
            最终答案：转换后的代码
            """),
            MessagesPlaceholder(variable_name="chat_history"),
            ("human", "{input}"),
            MessagesPlaceholder(variable_name="agent_scratchpad")
        ])
        
        # 创建Agent
        self.agent = create_structured_chat_agent(
            llm=self.llm,
            tools=self.tools,
            prompt=self.prompt
        )
        
        # 创建Agent执行器
        self.agent_executor = AgentExecutor.from_agent_and_tools(
            agent=self.agent,
            tools=self.tools,
            memory=self.memory,
            verbose=True
        )
        
    def load_rules(self, source_lang: str, target_lang: str) -> Dict:
        rules_file = os.path.join("rules", f"{source_lang}_to_{target_lang}.json")
        if os.path.exists(rules_file):
            try:
                with open(rules_file, 'r', encoding='utf-8') as f:
                    return json.load(f)
            except Exception as e:
                print(f"警告：加载规则文件时出错: {str(e)}")
        return {}
        
    def convert_code(self, code: str, source_lang: str, target_lang: str) -> str:
        # 加载转换规则
        rules = self.load_rules(source_lang, target_lang)
        
        # 构建输入
        input_text = f"""请将以下{source_lang}代码转换为{target_lang}代码：
        {code}
        """
        
        # 执行转换
        response = self.agent_executor.invoke({
            "input": input_text,
            "rules": json.dumps(rules, indent=2)
        })
        
        return response["output"]

def main():
    # 创建转换器实例
    transformer = LangChainTransformer()
    
    # 示例代码
    source_code = """
    import sys
    
    class Calculator:
        def add(self, a, b):
            return a + b
            
    def main():
        calc = Calculator()
        print(calc.add(1, 2))
        
    if __name__ == "__main__":
        main()
    """
    
    # 转换代码
    converted_code = transformer.convert_code(
        code=source_code,
        source_lang="Python",
        target_lang="Java"
    )
    
    # 保存结果
    with open("output.java", "w", encoding="utf-8") as f:
        f.write(converted_code)

if __name__ == "__main__":
    main() 