git_mulagent/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
python-dotenv>=1.0.0
openai>=1.12.0
autogen>=0.2.0
pyautogen>=0.2.0
ollama>=0.1.6
fix-busted-json>=0.0.1
python-pptx>=0.6.21 